move herd label formatting to separate function
